Improved the way the application is configured.
Administrator-relevant config is now put in a properties file with the same name as the .war in the webapps dir. The application will report it if a required setting is missing. All other configuration is in the .war file. The Ant build will copy the relevant files for the project variant we're building. The default Ant target "all" builds all project variants.