Additional BLS-port related changes.
* Eliminated searchjobresults.xsl; application uses the results template
right away and the template determines what HTML/JS code is produced
based on what BLS answers (status message or results). This also means
the extra "getResults" parameter is not added to the URL to get the
results; the page is simply reloaded. In the future we will combine the
result XSLTs into a single one that simply produces the appropriate
results page based on the BLS response XML, simplifying things even
further.
* Made "check again time" dynamic based on BLS's advice.
* Added Medium field to CHN search.xml.