Continue extracting corpus-specific content.
Move a lot of static content around to proper locations.
Move xslt files to the default interface dir under WEB-INF just like
custom corpora.
Move (fav)icons to /img/ .
Extract generated css (for background image and styling) to a separate
template.
Remove the defaultCorpus stuff that determines what version of pages to
show and what title to set. More improvements will be made in future
commits.
Expand BaseResponse with the corpus the response is being generated for,
and add a setting that configures whether the response can also be
generated without the corpus being set (for the about response for
instance).
Rewrite MainServlet:processRequest to remove the need for mapping
responses to /page/ . This also fixes the issue where missing trailing
slashes would break relative paths to static resources.