Work on the xpath building within the form.
Every field has a an internal xpathContext which describes how the
expression for that field was formed. The fields create the final
expression by concatenating all these contexts and creating an
expression. When opening the tree modal for a field, the xpath for that
field is evaluated in the background and nodes are highlighted.