Add support for "within" clause to simple search
Fix issue where qson would output strings containing the ' quote, which
is url-escaped in chrome. This fixes an issue where forward history
would be deleted by a pushState call when navigating back on the page.
This happened because a search is executed when navigating back, which
in turn causes a new querystring to be generated. This string is then
compared to the actual querystring in the navigation bar, and if it's
determined to be different, a pushState is performed to update the bar.
In this case the generated querystring contained the ' quotes, while the
navigation bar contained their url-escaped counterparts, causing an
erroneous pushState.