Fix bug where filter fields would refresh search after first submit
After a search is submitted, the store has two instances of the active filter fields.
One is bound to the interface, the other is just a cached version so that changes to filters in
the ui do not change results until the form is submitted.
The issue was that these could become the same pointer, causing ui changes to also update the
"cached" copy, causing it to update the results anyway...