Fix random sample seed being used for every search request
Blacklab generates a random sample seed when "samplenum" or "sample" parameters
are passed without the "sampleseed" parameter.
We allowed this, but a result of this was that every time the user viewed a different
page of results, a new seed was used, so the results were non-deterministic.
This causes issues with large grouped result sets, where the number of groups
changed every time you switched page.

Now we forcibly generate a seed whenever a sample size is also used.

- fixed some potential bugs where strings could end up in number fields in the store