Add support for gap-filling in the expert view
Like cql patterns, long lists of gap-filled values will be removed from the page url.
Submitting a query with different gap values will result in a new history entry in both the history view, and in the browser itself.

- Fixed a nasty bug where a previously submitted query would remain in the background when the page was reset.
This query would then "magically" reappear in the form when the user went back and then forward again in the browser.
- Improved documentation for url and history generation

Issue #157