Fix hidden annotations being parsed from query
This could happen when using the querybuilder or expert view and
entering values for annotations that aren't in the extended search.

In this case, refreshing the page would decode this query into hidden fields
and when the search was submitted from "extended", you'd see unintuitive results.
Additionally there was also no way to clear these values without resetting everything

Issue #279