Add option to display hits as html, and configure the displayed annotation
This commit adds the option to configure which annotation in shown
in the hit columns, as well as what annotation is used to display detailed snippets.

This is useful in itself to allow showing a different annotation than is configured as "main" annotation.
For instance for diacritic words.
In diacritic words, there are usually two words, but only one lemma shared between them.
Indexing multiple words per token is supported by blacklab,
but due to a limitation in the forward index, only the first value for an annotation can be shown in the results.
Using this system you can use a different annotation for searching (with multiple indexed values)
than you use for displaying the results (with the multiple words concatenated for example).

Additionally, you can now toggle to display the returned results as html.
This can be useful for when your words contain inline data, such a strike throughs.
You can index the words' raw xml into a separate annotation, and use it for displaying,
and set to display it as html. The inline tags can then be styled as normal using CSS.

Browser support for unknown tags is good, so arbitrary xml should work out of the box,
given it is well-formed.

USE THIS FEATURE WITH CARE! This is essentially a way to inject arbitrary scripts and data into the page.