Update dependencies and compiler toolchain
This also fixes a bunch of type checking errors.
Due to the addition of the types: [] property in tsconfig.json, not all type definitions present in node_modules are autoloaded anymore. This behavior was an issue because some of our dependencies pull in other dependencies that polluted the global type scope (during compilation/linting/autocompletion in vscode).
These types could be very broad (the entirity of the nodejs core library.. ugh) leading to several type name conflicts where the typescript compiler would think we were using one type, but runtime it would be the other type (due to using nodejs types, but actually running code in the browser -  a different environment).