Add ability to set pagination per-corpus
Defaults to 1000 words now.
The global setting has been removed, there is no longer a hard cap, though it has to be configured very high per corpus.

Pagination on the clientside should now also always be in sync with the server. We sanitize the pagination on the server and then send the sanitized version back to the client.

Also use Optional instead of null in a lot of places in the java code now.

Issue #331
Issue #330