Update node version for compilation to latest
Because node-sass needs to be compiled against node, this needs to match the version of node installed on a development machine (if you're also running dev builds).
We'd rather require the most recent version of node than force a developer to use an older one.

NOTE; you can work around this by uninstalling node-sass prior to building a release. The build process will then reinstall it against the node version specified in pom.xml. After the build you'll have to reinstall node-sass again to change it back to the system's version of node.