Fix requests not being cancelled when invalid characters are typed
For example when the input is 'a' and a search is already running, typing a space before the results come in would not cancel the request for results for - the now stale - search 'a'.
When the old request would complete you'd see suggestions for 'a' appear, despite the current input being something unactionable.

Issue #336