Fix issue where range queries were restored incorrectly after refresh
In the case of a multi-field query (year from/to) for example, the query was decoded into the combined filter, as well as the individual filters. Then when the search would be executed both those filters would execute, creating a query that would never return results.

Issue https://github.com/INL/corpus-frontend-config/issues/82