Fix for the fix to the bug where range filters would "double up"
Explanation:

Filter-range-multiple-fields would encode a filter as being fieldA:[x to y] AND fieldB:[x to y]. All good. Now when the page is refreshed, this needs to be decoded from the url again.
However this query actually matches 3 filters, the filter for fieldA, the filter for fieldB, and the special filter-range-multiple-fields for the combination of those 2 fields.
So the values are decoded and placed back into those 3(!) filters. Now when the query is then submitted again (to restore the results), those 3 filters are put in the query, which ends up being something like fieldA:[x to y] AND fieldB:[x to y] AND fieldA:(x or y) AND fieldB:(x or y). Clearly not what was intended!

The solution in this commit?
While decoding the values back into all the filters, do this in a specific order:
- First process the "special" filters. The range filter then deletes the decoded values for fieldA and fieldB.
- Only then are the normal filters processed. Now the individual filters for fieldA and fieldB will no longer see the decoded values (as those were consumed by the combination filter), and thus stay empty.
This way the value no longer "doubles up", and we can avoid hacks with marking fields as non-restoreable, or other such trickery.

Issue #379