Exclude filters in current tab from being counted in other tabs as well
When the filter Author is present in the currently selected filter tab, the tab will show a pill with (1) to indicate 1 active filter. If the Author filter is also in another tab, it will not be counted there anymore. This is to prevent the illusion that multiple filters are active when in fact it's just multiple instances of the same filter.

Issue INL/corpus-frontend-config#89