Works!
Not ideal yet, we use a jackson implementation for json, and a separate jaxb implementation for xml, meaning we have to add both sets of @annotations on our POJO to configure the mapping to json and xml respectively.

It's possible to use jackson as XML provider (instead of using jaxb) as well, using the jackson-dataformat-xml dependency, but while this will work, jersey won't know about it, so will complain there's no xml objectmapper.

We'll have to find a solution to that, perhaps writing our own bit of boilerplate (in Setup) to connect the two together, and register jackson as xml provider in addition to json