Add config option "withCredentials" (default false)
Setting this to true will pass authentication cookies to blacklab. This is almost never required, but might be needed if you want to develop against a secured instance of BlackLab. In that case, the corpus-frontend would run on localhost, but BlackLab is already deployed somewhere. By default authentication headers/cookies aren't passed along, which would make this setup impossible. But using this setting, it can be made possible.