Add option to configure clientside root path, for proxy situations.
Set the option 'cfUrlexternal' in corpus-frontend.properties.
This url will be the base url relative to which all css, js, and inter-page links are generated. It might have to be different than the paths on the host server sometimes, for instance when there is a proxy that rewrites /corpus-frontend/... to /something-else/... (perhaps when hosting multiple versions, or when hosting multiple dockerized instances on the same server, which all want to map to /corpus-frontend).

Also improve clientside url generation a little. It should be more robust in exotic deployments (such as on root, or when preceded by multiple path segments).