Fix document statistics no longer working
Added a workaround for the issue where BlackLab 4 snippet and hit responses have a slightly different shape when there's no context information.
Normally we have {left, match, right} properties, and previously this would always be true, for docs, hits and snippets. Since BlackLab 4, {left, right} are omitted when there's no wordsAroundHit (or when the hit lies right at the document boundary). Additionally, {match} was replaced by {snippet} in /snippet requests.
We now map those responses back to the original object shape. This will only be until we migrate to blacklab api v5, but that won't be until after an official release of v4.