Refactor results components
Split results view into more separate components that don't use the vuex store.
Behavior is identical, but it should now be easier to implement custom result views
without having to reimplement a lot of the smaller components.

A next step would be to pull up the breadcrumbs, totals, groupby, pagination, sort and export
components up into ResultsView itself, instead of passing them down the slots.
Then we could remove some more duplicate code from the actual results coponents.
After that, we could modify the custom view system to not override the ResultsView, but instead only override the GroupResults/HitResults/DocResults component.

Probably something like:
Results.vue renders a ResultsView with :showSort, :showGroup, :show... etc
ResultsView does the rendering of those components, and the wiring together.
It determines which component to render for the actual result displaying, or it uses the CustomView component name if it it's been set.