Re-add basic auth forwarding from frontend to blacklab
It's useful for local development when the frontend is not behind an auth layer, but blacklab is.

In that case, the backend of the corpus-frontend can't reach blacklab-server, unless it has the proper authorization header value, but it won't have that, because the interface is not behind the basic auth protection, so the browser won't prompt.

This way, the frontend will prompt whenever blacklab does