Fix a long wait before page shows a spinner if blacklab is slow
We need to know if there's a user before we do _anything_, as otherwise requests might fail.
If there's no oidc configured, we would try to contact BlackLab to see if that knows about a user, but that could take a long time.
That could potentially take a long time, so don't wait for that before rendering the page. (as that user is implicit and purely cosmetic).