Fix validation behavior of ui configuration
The old implementation would never actually do any validating...

Customjs would set some initial settings, which would be (temporarily) stored in the initialState, because the corpus wasn't loaded yet and the actual store instance not created yet.

Then when the init() function runs, the initialState was copied into the actual store instance. We would (badly) attempt to validate that.
The issue is that when we run a setter (actions.*) that does some internal validation, however, if the validation fails, the set is not applied. Which is not what we want, as we're trying to ferret out invalid elements in the store, so if the setter doesn't apply, the invalid value remains!

Fixed this by using a hackish workaround that makes the setter clear values during init() if the new value doesn't pass validation.
So now it's setsomevalue(getsomevalue()) which clears the somevalue if it's invalid.
Then we check afterwards if the value is empty, and replace it with the default if it is.