Attempt to use antlr4 parsing to provide syntax highlighting and code completion in the expert cql editor.
It seems the plumbing mostly works, i.e. we get tokens and syntax highlighting and autocompletion, the results are just garbage.
It seems the tokenization is written in such a way that it cannot handle incomplete queries well. And we need much more disambiguation between tokens, such as wel something is the default token (e.g. "word") vs when it's the value of an attribute [word="word"].
There are also compound tokens, such as the arrow operators, that prevent highlighting the infix dependency filter differently. We can fixup some of that stuff in a stage after the tokenizer, but it doesn't make it easy.
Autocompletion also outputs the name of expected tokens at the moment, not their expected contents. So we need a bunch more plumbing there.