Remove invalid grouping criteria automatically
For example the query a:"the" and grouping on capture label "a". Then modifying the query to remove the "a" label. The grouping is still active, but now BlackLab returns an error.
Detect this, and remove the offending group, so the user can continue.

We need this logic because the grouping window is hidden when there are no results, so there is no other way for the user to remove the grouping criterium.