Re-add some interop code when connecting to older BlackLab
Not all features will work, but this way things will at least render.
This is mostly useful for development. In production this will probably only be confusing, as the app appears to work but some parts may fail unexpectedly.

When the next major of BlackLab eventually releases we should add some proper version bounds and a clear warning to the page when operating against a BlackLab version outside the explicitly supported version(s).