Parallel: show correct highlighting for doc version shown.
When highlighting document contents in a parallel corpus, we
have to make sure the hits in the correct version are used for
highlighting.

This uses BlackLab's new BCQL function rfield() and parameter
searchfield to achieve this.

E.g. BLS/CORPNAME/docs/ID/contents?field=nl&searchfield=en&patt=rfield("dog" ==>nl _, 'nl')

Above example will show the Dutch version of a document with translations
for the word "dog" highlighted. This is done with a query that will search the
en field for dog and find alignments in the nl field, then will only return those
nl hits.

Squashed commit of the following:

commit 34c995230c95ac49119c37dc67c4b4f952b5591d
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Mon Aug 19 15:36:46 2024 +0200

    Parallel: show correct highlighting for doc version shown.

    When highlighting document contents in a parallel corpus, we
    have to make sure the hits in the correct version are used for
    highlighting.

    This uses BlackLab's new BCQL function rfield() and parameter
    searchfield to achieve this.

    E.g. BLS/CORPNAME/docs/ID/contents?field=nl&searchfield=en&patt=rfield("dog" ==>nl _, 'nl')

    Above example will show the Dutch version of a document with translations
    for the word "dog" highlighted. This is done with a query that will search the
    en field for dog and find alignments in the nl field, then will only return those
    nl hits.

commit 8fc1693357b6d8f2618ba325f125f10c1cf4c56b
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Mon Aug 5 15:57:03 2024 +0200

    Prepare to fix parallel doc highlighting.

commit 92586b4ccf7fbbded3d940fd8284ab93188b8b02
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Mon Aug 5 15:56:16 2024 +0200

    Hide docInfo.tokenCounts on metadata page.