Improvements to linting & remove unused dependencies
This also possibly fixes an issue with unknown namespace 'Node' warnings/errors in various files. We don't want to have 'node' types in our components, but that means that node is not a part of the global environment, which makes typescript think all files that depend on nodejs (such as webpack configs and such) are broken. So I've added those files to the tsconfig's exclude list to make typescript ignore them.

Linting:
Since the webpack vue-template-compiler/vue-plugin doesn't report errors in templates, add an automatic background task to vscode that does that.
It runs vue-tsc, which does report those errors (like undefined variables/functions used in templates), and prints the errors to the problems view. This has the added benefit that it processes the entire project instead of just the open files.
The task should autostart, but if it's annoying, you can kill the task and live without error reporting. Ignorance is bliss...

An improvement might be to use eslint instead of vue-tsc, so we can add warnings about unused variables and use of 'any' etc, but that would still need to run vue-tsc under the hood to properly work. Don't want to go that far for now.