Fix uid mixin breaking typescript a little bit
Using ANY mixin in vue 2 apparently makes it so that using an unknown property on the component no longer gives an error. Instead an unknown property simple becomes 'any'. For obvious reasons that's not great. You won't get an error when renaming a prop that's being used elsewhere for example.